backend/user_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
requests